registry1_2.uew ---------------
This wordfile is like registry1_1.uew with the difference of specifying
also the comma as delimiter character. Therefore just the hexadecimal
bytes are highlighted with the assigned color of color groups "Numbers"
and "Hex. Bytes", but not the commas and the backslashes at end of lines.

It is recommended to assign the same color to color group "Numbers" and
"Hex. Bytes" to get identical syntax highlighting for hexadecimal values
starting with [0-9a-f].