merge registry1_3.uew ---------------
This wordfile is like registry1_2.uew with the difference of no color group
"Hex. Bytes". The color of color group "Numbers" should be set to the color
of normal text when using this wordfile to get all the hexadecimal bytes
highlighted as normal text. This speeds up the syntax highlighting a little
bit as the hexadecimal bytes are the only strings not special highlighted.
The commas and backslashes are highlighted instead.