merge registry2_1.uew ---------------
This wordfile does not support code folding. It interprets lines starting
with [ as comment line. Customize the color group "Comments" to adjust the
color for [keys]. In comparison to registry1_1.uew the syntax highlighting
with this wordfile is slightly faster which makes it better for syntax
highlighting *.reg files with several MB.

This wordfile highlights hexadecimal bytes, the commas between the bytes
and the backslashes at end of lines with hexadecimal bytes in the color
assigned to color group "Numbers" and "Hex. Bytes".

It is recommended to assign the same color to color group "Numbers" and
"Hex. Bytes" to get identical syntax highlighting for hexadecimal values
starting with [0-9a-f].